% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_complications_number}
\alias{pasteur_complications_number}
\title{pasteur_complications_number}
\usage{
pasteur_complications_number()
}
\description{
complications_number function that produces the number of complications for each study_id
}
\examples{
\dontrun{
pasteur_complications_number()
}
}
