% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_consented}
\alias{pasteur_consented}
\title{pasteur consented}
\usage{
pasteur_consented()
}
\description{
this column is determined by checking crf00 for whether the participant consented in both locations in the data
}
\examples{
\dontrun{
pasteur_consented()
}
}
