% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_definitive_fixation_date}
\alias{pasteur_definitive_fixation_date}
\title{pasteur definitive_fixation_date}
\usage{
pasteur_definitive_fixation_date()
}
\description{
definitive_fixation_date function gets the surgery data from the CRF07 Definitive Fixation Form
}
\examples{
\dontrun{
pasteur_definitive_fixation_date()
}
}
