% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_early_withdraw}
\alias{pasteur_early_withdraw}
\title{pasteur early_withdraw}
\usage{
pasteur_early_withdraw()
}
\description{
this column is determined by reviewing the early_withdraw_reason column and if there is a reason then there was an early withdraw
}
\examples{
\dontrun{
pasteur_early_withdraw()
}
}
