% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_early_withdraw_reason}
\alias{pasteur_early_withdraw_reason}
\title{pasteur early_withdraw_reason}
\usage{
pasteur_early_withdraw_reason()
}
\description{
this column is determined by checking af00 for whether the participant discontinued in the study
}
\examples{
\dontrun{
pasteur_early_withdraw_reason()
}
}
