% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_eligible}
\alias{pasteur_eligible}
\title{pasteur eligible}
\usage{
pasteur_eligible()
}
\description{
this column is determined by checking crf00 for whether the participant is eligible for the study
}
\examples{
\dontrun{
pasteur_eligible()
}
}
