% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_enrolled}
\alias{pasteur_enrolled}
\title{pasteur enrolled}
\usage{
pasteur_enrolled()
}
\description{
this column is determined by comparing the randomized to the early_withdraw columns in the analytic set
}
\examples{
\dontrun{
pasteur_enrolled()
}
}
