% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_enrolled_consider_covid}
\alias{pasteur_enrolled_consider_covid}
\title{pasteur enrolled_consider_covid}
\usage{
pasteur_enrolled_consider_covid()
}
\description{
enrolled function imports the consented, adjudicated_ineligible, pending_adjudication, 
has_early_discontinuation, and not_added_during_covid and combines them into a boolean for whether each ID was
enrolled excluding those enrolled during the height of covid.
}
\examples{
\dontrun{
pasteur_enrolled_consider_covid()
}
}
