% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_injury_type}
\alias{pasteur_injury_type}
\title{pasteur injury_type}
\usage{
pasteur_injury_type()
}
\description{
injury type is determined by checking crf00 for the fracture types and recording ankle or plateau for each
}
\examples{
\dontrun{
pasteur_injury_type()
}
}
