% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_not_consented}
\alias{pasteur_not_consented}
\title{pasteur not_consented}
\usage{
pasteur_not_consented()
}
\description{
this column is determined by comparing the refused, eligible, and consented columns in the analytic set
}
\examples{
\dontrun{
pasteur_not_consented()
}
}
