% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_pending_verification}
\alias{pasteur_pending_verification}
\title{pasteur pending_verification}
\usage{
pasteur_pending_verification()
}
\description{
this column is determined by checking af00 for whether the participant discontinued in the study
}
\examples{
\dontrun{
pasteur_pending_verification()
}
}
