% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_randomized}
\alias{pasteur_randomized}
\title{pasteur randomized}
\usage{
pasteur_randomized()
}
\description{
this column is determined by checking multiple crfs for whether 
the participant was actually randomized into the study
}
\examples{
\dontrun{
pasteur_randomized()
}
}
