% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_refused}
\alias{pasteur_refused}
\title{pasteur refused}
\usage{
pasteur_refused()
}
\description{
this column is determined by checking crf00 for whether the participant refused to consent or not
}
\examples{
\dontrun{
pasteur_refused()
}
}
