% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_six_month_radiographs_taken}
\alias{pasteur_six_month_radiographs_taken}
\title{pasteur six_month_radiographs_taken}
\usage{
pasteur_six_month_radiographs_taken()
}
\description{
six_month_radiographs_taken function imports the injury_type construct, considers only the ankle injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_six_month_radiographs_taken()
}
}
