% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_six_week_ankle_sagital_displacement}
\alias{pasteur_six_week_ankle_sagital_displacement}
\title{pasteur six_week_ankle_sagital_displacement}
\usage{
pasteur_six_week_ankle_sagital_displacement()
}
\description{
six_week_ankle_sagital_displacement function imports the injury_type construct, considers only the ankle injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_six_week_ankle_sagital_displacement()
}
}
