% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_six_week_plateau_angle_between_femur_and_tibia}
\alias{pasteur_six_week_plateau_angle_between_femur_and_tibia}
\title{pasteur six_week_plateau_angle_between_femur_and_tibia}
\usage{
pasteur_six_week_plateau_angle_between_femur_and_tibia()
}
\description{
six_week_plateau_angle_between_femur_and_tibia function imports the injury_type construct, considers only the plateau injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_six_week_plateau_angle_between_femur_and_tibia()
}
}
