% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_six_week_plateau_medial_proximal_tibia_angle}
\alias{pasteur_six_week_plateau_medial_proximal_tibia_angle}
\title{pasteur six_week_plateau_medial_proximal_tibia_angle}
\usage{
pasteur_six_week_plateau_medial_proximal_tibia_angle()
}
\description{
six_week_plateau_medial_proximal_tibia_angle function imports the injury_type construct, considers only the plateau injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_six_week_plateau_medial_proximal_tibia_angle()
}
}
