% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_three_month_ankle_coronal_plane_displacement}
\alias{pasteur_three_month_ankle_coronal_plane_displacement}
\title{pasteur three_month_ankle_coronal_plane_displacement}
\usage{
pasteur_three_month_ankle_coronal_plane_displacement()
}
\description{
three_month_ankle_coronal_plane_displacement function imports the injury_type construct, considers only the ankle injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_three_month_ankle_coronal_plane_displacement()
}
}
