% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_twelve_month_plateau_articular_step_off_lateral}
\alias{pasteur_twelve_month_plateau_articular_step_off_lateral}
\title{pasteur twelve_month_plateau_articular_step_off_lateral}
\usage{
pasteur_twelve_month_plateau_articular_step_off_lateral()
}
\description{
twelve_month_plateau_articular_step_off_lateral function imports the injury_type construct, considers only the plateau injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_twelve_month_plateau_articular_step_off_lateral()
}
}
