% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_twelve_month_plateau_sagittal_plane_alignment}
\alias{pasteur_twelve_month_plateau_sagittal_plane_alignment}
\title{pasteur twelve_month_plateau_sagittal_plane_alignment}
\usage{
pasteur_twelve_month_plateau_sagittal_plane_alignment()
}
\description{
twelve_month_plateau_sagittal_plane_alignment function imports the injury_type construct, considers only the plateau injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_twelve_month_plateau_sagittal_plane_alignment()
}
}
