% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteur_code.R
\name{pasteur_twelve_month_radiographs_taken}
\alias{pasteur_twelve_month_radiographs_taken}
\title{pasteur twelve_month_radiographs_taken}
\usage{
pasteur_twelve_month_radiographs_taken()
}
\description{
twelve_month_radiographs_taken function imports the injury_type construct, considers only the ankle injury participants, and then looks at a data point from a particular field in the followup from
}
\examples{
\dontrun{
pasteur_twelve_month_radiographs_taken()
}
}
