% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_baseline.R
\name{snow_facilitycode}
\alias{snow_facilitycode}
\title{Snow facilitycode}
\usage{
snow_facilitycode()
}
\value{
standard constructor output
}
\description{
To determine facilitycode, we just need to consider the screening form
and for that we only need data from the baseline event.
}
\examples{
\dontrun{
snow_facilitycode()
}
}
