% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_baseline.R
\name{snow_screened}
\alias{snow_screened}
\title{Screened}
\usage{
snow_screened()
}
\value{
standard constructor output
}
\description{
To determine if a participant is screened we just need to consider if the CRF00 is complete
and for that we only need data from the baseline event.
}
\examples{
\dontrun{
snow_screened()
}
}
